Allow forcing SPT version in entrypoint via FORCE_SPT_VERSION env var

+++ entrypoint.sh
@@ -10,9 +10,11 @@ backup_dir_name=${BACKUP_DIR:-backups}
 backup_dir=$mounted_dir/$backup_dir_name
 
 spt_current_major_version=4
-spt_version=${SPT_VERSION:-4.0.0-40087-0582f8d}
+spt_version=${SPT_VERSION:-4.0.1-40087-1eacf0f}
 spt_version=$(echo $spt_version | cut -d '-' -f 1)
 spt_backup_dir=$backup_dir/spt/$(date +%Y%m%dT%H%M)
+force_spt_version=${FORCE_SPT_VERSION:=}
+forced_spt_version_archive=SPT-${force_spt_version}.yz
 
 nodejs_spt_data_dir=$mounted_dir/SPT_Data
 spt_nodejs_core_config=$nodejs_spt_data_dir/Server/configs/core.json
@@ -20,7 +22,8 @@ spt_dir=$mounted_dir/SPT
 spt_data_dir=$spt_dir/SPT_Data
 enable_spt_listen_on_all_networks=${LISTEN_ALL_NETWORKS:-false}
 
-fika_version=${FIKA_VERSION:-1.0.1}
+
+fika_version=${FIKA_VERSION:-1.0.2}
 install_fika=${INSTALL_FIKA:-false}
 fika_backup_dir=$backup_dir/fika/$(date +%Y%m%dT%H%M)
 fika_config_path=assets/configs/fika.jsonc
@@ -228,7 +231,18 @@ set_num_headless_profiles() {
 install_spt() {
     # Remove the server files, since databases tend to be different between versions
     rm -rf $spt_data_dir
-    cp -r $build_dir/* $mounted_dir
+
+    # If FORCE_SPT_VERSION is set, download and override the built in version with provided version
+    # Archive stored in root mounted folder. Supports user manually supplying the release archive
+    if [[ -n ${force_spt_version} ]]; then
+        # check if archive already exists, and extract if so
+        if [[ ! -f ${build_dir}/${forced_spt_version_archive} ]]; then
+            curl -sL "https://spt-releases.modd.in/SPT-${force_spt_version}.7z" -o SPT-${force_spt_version}.7z
+        fi
+        7zz x ${forced_spt_version_archive}
+    else
+        cp -r $build_dir/* $mounted_dir
+    fi
     make_and_own_spt_dirs
 }
